+++ cmd/musictui/types.go
@@ -20,8 +20,8 @@ type PlayerState struct {
 	Type            string                 `json:"type"`
 	IsPlaying       bool                   `json:"is_playing"`
 	CurrentTrack    *types.Track           `json:"current_track"`
-	CurrentTime     int                    `json:"current_time"`
-	Duration        int                    `json:"duration"`
+	CurrentTime     float64                `json:"current_time"`
+	Duration        float64                `json:"duration"`
 	Volume          float64                `json:"volume"`
 	Muted           bool                   `json:"muted"`
 	Shuffle         bool                   `json:"shuffle"`
